+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 
 * Removal: `Scalingo::Client#agora_fr1` had been removed since the region no longer exists.
+* New: Add `addons#authenticate!` endpoint
+* New API: database API
+* New API: backup API
 
 ## 3.1.0